Multithreading
Description: Project 2 has the objective of sorting an array of 32-bit integers using the Quick sort, Insertion sort, and Shell sort (Hibbard's Sequence) algorithms. There are a number of arguments the user can provide (some optional and some not) which determine how the program will process the sorting. We are allowed to run the sorting on up to 4 threads or less which is determined by the user.